+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/visualmode/VisualMode.java
@@ -531,14 +531,6 @@ public class VisualMode implements VisualModeEditorSync,
                         allDone.execute(false);
                         return;
                      }
-                                          
-                     // visual editor source capsules get tripped up on verbatim chunks
-                     if(hasVerbatimChunk(result.canonical))
-                     {
-                        view_.showWarningBar(constants_.unableToActivateVisualEditorVerbatim());
-                        allDone.execute(false);
-                        return;
-                     }
                      
                      // if we failed to extract a source capsule then don't switch (as the user will have lost data)
                      if (hasSourceCapsule(result.canonical))
@@ -1835,11 +1827,6 @@ public class VisualMode implements VisualModeEditorSync,
       // (note that this constant is also defined in rmd_chunk-capsule.ts)
       final String kRmdBlockCapsuleType = "f3175f2a-e8a0-4436-be12-b33925b6d220".toLowerCase();
       return markdown.contains(kRmdBlockCapsuleType);
-   }
-   
-   private boolean hasVerbatimChunk(String markdown)
-   {
-      return markdown.contains("\n```{verbatim}");
    }
    
    /**